\name{checkModelFit}
\alias{checkModelFit}
\title{Calibration of aquifer parameter "hkr"}
\description{Calibration of aquifer parameter "hkr"}
\usage{checkModelFit(wtaqResult, wells = ".*")}
\arguments{
  \item{wtaqResult}{# in WTAQ result in "long format"}
  \item{wells}{regular expression of wells to be included (e.g. ".*" for all, PW: only
production well, OW: all observation wells)}
}



\author{Hauke Sonnenberg}





