\name{drawdowns}
\docType{data}
\alias{drawdowns}
\title{Example data of measured drawdowns in a drinking water well field}
\description{
Example data of measured drawdowns in a drinking water well field.
}
\usage{drawdowns}
\format{List of five data frames each of which represents the timeseries of
drawdowns measured at five different drinking water wells. The column 
\emph{time.in.seconds} is the time of operation in seconds and the columns 
\emph{W1}, \emph{W2}, \emph{W3}, \emph{W4}, and \emph{W5} correspond to the
drawdowns at the different Wells W1 to W5.}
\keyword{drawdowns}