\name{fitnessAquifer}
\alias{fitnessAquifer}
\title{fitnessAquifer}
\description{fitnessAquifer}
\usage{fitnessAquifer(parameterValue, parameterName, configuration, 
    wellPattern = "*")}
\arguments{
  \item{parameterValue}{
}
  \item{parameterName}{
}
  \item{configuration}{
}
  \item{wellPattern}{
}
}



\author{Hauke Sonnenberg}





