\name{kwb.wtaq-package}
\alias{kwb.wtaq-package}
\alias{kwb.wtaq}
\docType{package}
\title{Interface to WTAQ Drawdown Model (http://water.usgs.gov/ogw/wtaq/)}
\description{Functions enabling the writing of WTAQ input files, running of WTAQ
and reading of WTAQ output files.}
\details{
\tabular{ll}{Package: \tab kwb.wtaq\cr
Maintainer: \tab Hauke Sonnenberg <hauke.sonnenberg@kompetenz-wasser.de>\cr
Depends: \tab lattice, hydroGOF, plotrix, reshape\cr
Remotes: \tab KWB-R/kwb.utils\cr
Author: \tab Hauke Sonnenberg\cr
Version: \tab 0.2.99\cr
License: \tab MIT + file LICENSE\cr
Title: \tab Interface to WTAQ Drawdown Model (http://water.usgs.gov/ogw/wtaq/)\cr}
}
\author{Hauke Sonnenberg}

\keyword{ package }


