\name{modelFitness}
\alias{modelFitness}
\title{modelFitness}
\description{modelFitness}
\usage{modelFitness(wtaqResult, wells = "*")}
\arguments{
  \item{wtaqResult}{# in "long format"}
  \item{wells}{regular expression of wells to be included (e.g. * for all, PW: only
production well, OW: all observation wells)}
}



\author{Hauke Sonnenberg}





