\name{modelFitnessAggregated}
\alias{modelFitnessAggregated}
\title{modelFitnessAggregated}
\description{modelFitnessAggregated}
\usage{modelFitnessAggregated(wtaqResult, wellPattern = "*")}
\arguments{
  \item{wtaqResult}{
}
  \item{wellPattern}{
}
}



\author{Hauke Sonnenberg}





