\name{owPlotDrawdowns}
\alias{owPlotDrawdowns}
\title{plot drawdown (under construction!)}
\description{plot drawdown (under construction!)}
\usage{owPlotDrawdowns(drawdownList, topview = TRUE, overlay = FALSE, 
    fixview = FALSE)}
\arguments{
  \item{drawdownList}{
}
  \item{topview}{
}
  \item{overlay}{
}
  \item{fixview}{
}
}



\author{Hauke Sonnenberg}





