\name{owSuperposeDrawdowns}
\alias{owSuperposeDrawdowns}
\title{superpose drawdowns}
\description{superpose drawdowns in list of drawdowns as returned by
\code{\link{owGetDrawdowns}}.}
\usage{owSuperposeDrawdowns(drawdownList, dbg = FALSE)}
\arguments{
  \item{drawdownList}{list of drawdowns as returned by \code{\link{owGetDrawdowns}}}
  \item{dbg}{if TRUE, debug messages are shown. Default: FALSE  }
}

\value{data frame}

\author{Hauke Sonnenberg}



\seealso{\code{\link{owGetDrawdowns}}      }

