\name{print.wtaqConfiguration}
\alias{print.wtaqConfiguration}
\title{Print WTAQ configuration}
\description{Print WTAQ configuration}
\usage{\method{print}{wtaqConfiguration}(x, ...)}
\arguments{
  \item{x}{
}
  \item{\dots}{
}
}



\author{Hauke Sonnenberg}





