\name{wtConfiguredWellnames}
\alias{wtConfiguredWellnames}
\title{wtConfiguredWellnames}
\description{Read all well names (pumping well \dQuote{PW} plus observation wells) from
WTAQ configuration \emph{configuration}}
\usage{wtConfiguredWellnames(configuration)}
\arguments{
  \item{configuration}{WTAQ configuration as generated by \code{\link{wtConfigure}}}
}

\value{vector of character containing the name of the pumping well (as the first
element) and the names of the observation wells as following elements
(according to the order of their definition in \emph{configuration}}

\author{Hauke Sonnenberg}





