\name{wtConfigureGeneral}
\alias{wtConfigureGeneral}
\title{General parameter values}
\description{Define general WTAQ parameters}
\usage{wtConfigureGeneral(title = paste("Input file generated by kwb.wtaq-functions on", 
    date()), format = "DIMENSIONAL")}
\arguments{
  \item{title}{Title of simulation; up to 70 characters in length. Leave this line blank 
if no title is specified.}
  \item{format}{Analysis format. Enter DIMENSIONAL.}
}

\value{list with elements \emph{title} and \emph{format}, representing general 
WTAQ parameters. See descriptions in Arguments section.}

\author{Hauke Sonnenberg}



\seealso{\code{\link{wtConfigure}}}

