\name{wtDefaultConfiguration}
\alias{wtDefaultConfiguration}
\title{Default WTAQ configuration}
\description{Default WTAQ configuration. }
\usage{wtDefaultConfiguration()}

\value{Currently: configuration for WTAQ sample problem 2, as also returned by
\code{\link{wtConfigurationExample2}}}

\author{Hauke Sonnenberg}



\seealso{\code{\link{wtConfigurationExample2}, \link{wtConfigurationExample3}}}

