\name{wtRepairConfiguration}
\alias{wtRepairConfiguration}
\title{repair configuration}
\description{repair configuration by resetting parameter values that depend on others}
\usage{wtRepairConfiguration(configuration, dbg = FALSE)}
\arguments{
  \item{configuration}{WTAQ configuration, as retrieved by \code{\link{wtConfigure}}.  }
  \item{dbg}{if TRUE, debug messages are shown, else not}
}



\author{Hauke Sonnenberg}





