\name{kwb.wtaq-package}
\alias{kwb.wtaq-package}
\alias{kwb.wtaq}
\docType{package}
\title{Interface to WTAQ Drawdown Model (http://water.usgs.gov/ogw/wtaq/)}
\description{Functions enabling the writing of WTAQ input files, running of WTAQ
and reading of WTAQ output files.}
\details{
\tabular{ll}{Package: \tab kwb.wtaq\cr
Title: \tab Interface to WTAQ Drawdown Model (http://water.usgs.gov/ogw/wtaq/)\cr
Author: \tab Hauke Sonnenberg\cr
Maintainer: \tab Hauke Sonnenberg <hauke.sonnenberg@kompetenz-wasser.de>\cr
Depends: \tab R (>= 3.0.0)\cr
Imports: \tab 
kwb.utils,
lattice,
hydroGOF,
plotrix,
reshape\cr
Suggests: \tab 
rmarkdown,
knitr\cr
Remotes: \tab github::KWB-R/kwb.utils\cr
BinaryFiles: \tab 
wtaq.2.1.exe,
wtaq.2.1_linux32bit.exe,
wtaq.2.1_linux64bit.exe\cr
Version: \tab 0.2.99\cr
License: \tab MIT + file LICENSE\cr
RoxygenNote: \tab 5.0.1\cr
VignetteBuilder: \tab knitr\cr}
}
\author{Hauke Sonnenberg}

\keyword{ package }


